Move src files into their new homes
