+++ lib/launch_core/dev/sensor.c
@@ -1,0 +1,55 @@
+#include <launch_core/dev/sensor.h>
+
+LOG_MODULE_REGISTER(launch_adc_utils);
+
+int l_update_get_sensor_data(const struct device *const dev, l_sensor_readings_args_t *args, bool convert_to_float) {
+    l_update_sensors(&dev, 1);
+
+    if (convert_to_float) {
+        l_get_sensor_data_float(dev, args->num_readings, args->channels, args->float_values);
+    } else {
+        l_get_sensor_data(dev, args->num_readings, args->channels, args->values);
+    }
+
+    return 0;
+}
+
+int l_update_sensors(const struct device *const *devs, int num_devs) {
+    for (int i = 0; i < num_devs; i++) {
+        int ret = sensor_sample_fetch(devs[i]);
+        if (ret != 0) {
+            LOG_ERR("Failed to fetch sensor data from %s. Errno %d\n", devs[i]->name, ret);
+        }
+    }
+
+    return 0;
+}
+
+int l_update_sensors_safe(const struct device *const *devs, int num_devs, const bool *devs_ready) {
+    for (int i = 0; i < num_devs; i++) {
+        if (unlikely(!devs_ready[i])) { // Skip if channel is not ready
+            LOG_ERR("Sensor %s is not ready.\n", devs[i]->name);
+            continue;
+        }
+
+        int ret = sensor_sample_fetch(devs[i]);
+        if (ret != 0) {
+            LOG_ERR("Failed to fetch sensor data from %s. Errno %d\n", devs[i]->name, ret);
+        }
+    }
+
+    return 0;
+}
+
+
+int l_get_sensor_data(const struct device *const dev, int num_channels, enum sensor_channel const *channels,
+                      struct sensor_value **values) {
+    for (int i = 0; i < num_channels; i++) {
+        int ret = sensor_channel_get(dev, channels[i], values[i]);
+        if (ret != 0) {
+            LOG_ERR("Failed to get sensor data from %s. Errno %d\n", dev->name, ret);
+        }
+    }
+
+    return 0;
+}
