+++ lib/launch_core/net/lora.c
@@ -4,11 +4,13 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
+#include <include/launch_core/net/lora.h>
 
 #include "zephyr/drivers/lora.h"
 #include <zephyr/kernel.h>
-#include <memory.h>
+#include <zephyr/logging/log.h>
 
+#include <memory.h>
 
 int l_lora_configure(const struct device *const dev, bool transmit) {
     struct lora_modem_config config = {
